Refactored Golang Code Organization and Corrected Import
